when number of entities is the same (greater 1), only use the one from existing coordinates file and don't complement
