+++ go.mod
@@ -1,10 +1,7 @@
 module github.com/osc-em/converter-OSCEM-to-mmCIF
 
-go 1.22.1
+go 1.23.0
 
-require (
-	github.com/BurntSushi/cif v0.0.0-20140118173227-e459106941ba // indirect
-	golang.org/x/sys v0.4.0 // indirect
-	golang.org/x/tools v0.5.1-0.20230111220935-a7f7db3f17fc // indirect
-	golang.org/x/tools/cmd/cover v0.1.0-deprecated // indirect
-)
+toolchain go1.23.8
+
+require golang.org/x/exp v0.0.0-20250408133849-7e4ce0ab07d0
