+++ parser/writeMmCif.go
@@ -4,19 +4,16 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
+	"io"
 	"log"
-	"math"
 	"os"
-	"regexp"
-	"sort"
-	"strconv"
 	"strings"
-	"time"
 
-	"github.com/osc-em/converter-OSCEM-to-mmCIF/converterUtils"
+	cU "github.com/osc-em/converter-OSCEM-to-mmCIF/converterUtils"
+	"golang.org/x/exp/slices"
 )
 
-func relevantId(PDBxItems map[string][]converterUtils.PDBxItem, dataItem converterUtils.PDBxItem) bool {
+func relevantId(PDBxItems map[string][]cU.PDBxItem, dataItem cU.PDBxItem) bool {
 	accessValues := PDBxItems[dataItem.CategoryID]
 	var parentValue string
 	for i := range len(accessValues) {
@@ -51,7 +48,7 @@ func getKeyByValue(value string, m map[string]string) (string, error) {
 }
 
 // given a slice of PDBx items get the length of a longest data item name in it (because the category is the same)
-func getLongestPDBxItem(s []converterUtils.PDBxItem, s2keys []string) int {
+func getLongestPDBxItem(s []cU.PDBxItem, s2keys []string) int {
 	var l int
 	for i := range s {
 		if len(s[i].Name) > l {
@@ -81,7 +78,9 @@ func mmCIFStringToPDBxItem(s string) (map[string][]string, []string, int) {
 			dqFields := (strings.Split(line, "\""))
 			if len(sqFields) != 1 {
 				for _, f := range sqFields {
-					if string(f[0]) == " " || string(f[len(f)-1]) == " " {
+					if f == "" {
+						continue
+					} else if string(f[0]) == " " || string(f[len(f)-1]) == " " {
 						fields = append(fields, strings.Fields(f)...)
 					} else {
 						fields = append(fields, f)
@@ -89,7 +88,9 @@ func mmCIFStringToPDBxItem(s string) (map[string][]string, []string, int) {
 				}
 			} else if len(dqFields) != 1 {
 				for _, f := range dqFields {
-					if string(f[0]) == " " || string(f[len(f)-1]) == " " {
+					if f == "" {
+						continue
+					} else if string(f[0]) == " " || string(f[len(f)-1]) == " " {
 						fields = append(fields, strings.Fields(f)...)
 					} else {
 						fields = append(fields, f)
@@ -139,272 +140,44 @@ func mmCIFStringToPDBxItem(s string) (map[string][]string, []string, int) {
 	return cifDI, keys, cifDIlen
 }
 
-func validateDateIsRFC3339(date string) string {
-	t, err := time.Parse(time.RFC3339, date)
-	if err != nil {
-		log.Printf("Date seems to be in the wrong format! We expect RFC3339 format, provided date %s is not.", date)
-		return ""
-	}
-	return t.Format(time.DateOnly)
-}
-
-func validateRange(value string, dataItem converterUtils.PDBxItem, unitOSCEM string, nameOSCEM string) (bool, error) {
-	rMin := dataItem.RangeMin
-	rMax := dataItem.RangeMax
-	unitPDBx := dataItem.Unit
-	namePDBx := dataItem.CategoryID + "." + dataItem.Name
-	var unitsSame bool
-	var errorMessage string
-	var unitsError error
-	if unitOSCEM == "" && unitPDBx == "" {
-		// both OSCEM and PDBx have no units definition
-		unitsSame = true
-	} else if unitOSCEM == "" && unitPDBx != "" {
-		errorMessage = fmt.Sprintf("No units defined for %s in OSCEM! Analogous property %s in PDBx has %s units. Value will still be used in mmCIF file!", nameOSCEM, namePDBx, unitPDBx)
-		unitsSame = false
-		unitsError = errors.New(errorMessage)
-		return true, unitsError
-	} else if unitOSCEM != "" && unitPDBx == "" {
-		errorMessage = fmt.Sprintf("No units defined for %s in PDBx! Analogous property %s in OSCEM has %s units. Value will still be used in mmCIF file!", namePDBx, nameOSCEM, unitOSCEM)
-		unitsSame = false
-		unitsError = errors.New(errorMessage)
-		return true, unitsError
-	} else {
-		explicitUnitOSCEM, ok := converterUtils.UnitsName[unitOSCEM]
-		if !ok {
-			errorMessage = fmt.Sprintf("No explicit unit name is specified for property %s in OSCEM, only a short name %s. Value will still be used in mmCIF file!", nameOSCEM, unitOSCEM)
-			unitsSame = false
-			unitsError = errors.New(errorMessage)
-			return true, unitsError
-		} else {
-			unitsError = nil
-		}
-		unitsSame = explicitUnitOSCEM == unitPDBx
-	}
-	// FIXME: when units are settled, do conversion when possible!
-	if unitsSame {
-		v, err := strconv.ParseFloat(value, 64)
-		if err != nil {
-			errorMessage = fmt.Sprintf("JSON value %s not numeric, but supposed to be", value)
-			return false, errors.New(errorMessage)
-		}
-		rMin, err := strconv.ParseFloat(rMin, 64)
-		if err != nil {
-			rMin = math.NaN()
-		}
-		rMax, err := strconv.ParseFloat(rMax, 64)
-		if err != nil {
-			rMax = math.NaN()
-		}
-		if math.IsNaN(rMin) && math.IsNaN(float64(rMax)) {
-			return true, unitsError
-		} else if math.IsNaN(rMin) {
-			return float64(v) <= rMax, unitsError
-		} else if math.IsNaN(rMax) {
-			return float64(v) >= rMin, unitsError
-		} else {
-			return float64(v) >= rMin && float64(v) <= rMax, unitsError
-		}
-	} else {
-		errorMessage = fmt.Sprintf("Units for analogous properties %s in OSCEM and %s in PDBx  don't match! Implement a converter from %s in OSCEM to %s expected by PDBx. Value will still be used in mmCIF file!", nameOSCEM, namePDBx, unitOSCEM, unitPDBx)
-		unitsError = errors.New(errorMessage)
-		return true, unitsError
-	}
-
-}
-func validateEnum(value string, dataItem converterUtils.PDBxItem) string {
-	enumFromEMDB := dataItem.EnumValues
-	enumFromPDBx := dataItem.PDBxEnumValues
-	namePDBx := dataItem.CategoryID + "." + dataItem.Name
-	if value == "true" {
-		return "YES"
-	} else if value == "false" {
-		return "NO"
-	}
-
-	if namePDBx == "_em_imaging.microscope_model" {
-		reTitan := regexp.MustCompile(`(?i)titan`)
-		if reTitan.MatchString(value) {
-			return "TFS KRIOS"
-		}
-	} else if namePDBx == "_em_imaging.mode" {
-		if value == "BrightField" {
-			return "BRIGHT FIELD"
-		}
-	} else if namePDBx == "_em_imaging.electron_source" {
-		if value == "FieldEmission" {
-			return "FIELD EMISSION GUN"
-		}
-	}
-	for i := range enumFromEMDB {
-		if strings.EqualFold(enumFromEMDB[i], value) {
-			value = enumFromEMDB[i]
-			return value
-		}
-	}
-	// scan through both enums
-	for i := range enumFromPDBx {
-		if strings.EqualFold(enumFromPDBx[i], value) {
-			value = enumFromPDBx[i]
-			return value
-		}
-	}
-	if namePDBx == "_em_imaging.illumination_mode" {
-		reFloodBeam := regexp.MustCompile(`(?i)parallel`)
-		if reFloodBeam.MatchString(value) {
-			return "FLOOD BEAM"
-		}
-	}
-	if namePDBx == "_pdbx_contact_author.role" {
-		reRole := regexp.MustCompile(`(?i)(principal investigator|group leader|pi)`)
-		if reRole.MatchString(value) {
-			return "principal investigator/group leader"
-		}
-	}
-	// add additional matching mechanism for grid material by chemical element name/ regular expression
-	if namePDBx == "_em_sample_support.grid_material" {
-
-		reGraphene := regexp.MustCompile(`(?i)graphene`)
-		reSilicon := regexp.MustCompile(`(?i)silicon`)
-		if reGraphene.MatchString(value) {
-			return "GRAPHENE OXIDE"
-		} else if reSilicon.MatchString(value) {
-			return "SILICON NITRIDE"
-		}
-		switch value {
-		case "Cu":
-			return "COPPER"
-		case "Cu/Pd":
-			return "COPPER/PALLADIUM"
-		case "Cu/Rh":
-			return "COPPER/RHODIUM"
-		case "Au":
-			return "GOLD"
-		case "Ni":
-			return "NICKEL"
-		case "Ni/Ti":
-			return "NICKEL/TITANIUM"
-		case "Pt":
-			return "PLATINUM"
-		case "W":
-			return "TUNGSTEN"
-		case "Ti":
-			return "TITANIUM"
-		case "Mo":
-			return "MOLYBDENUM"
-		}
-
-	} else if namePDBx == "_em_image_recording.film_or_detector_model" {
-		// add additional matching mechanism for "Falcon" detector model by regular expression; other don't seem feasible
-
-		reFalconI := regexp.MustCompile(`(?i)falcon[\s_]*?(1|I)`)
-		reFalconII := regexp.MustCompile(`(?i)falcon[\s_]*?(2|II)`)
-		reFalconIII := regexp.MustCompile(`(?i)falcon[\s_]*?(3|III)`)
-		reFalconIV := regexp.MustCompile(`(?i)falcon[\s_]*?(4|IV)`)
-		switch {
-		case reFalconIV.MatchString(value):
-			return "FEI FALCON IV (4k x 4k)"
-		case reFalconIII.MatchString(value):
-			return "FEI FALCON III (4k x 4k)"
-		case reFalconII.MatchString(value):
-			return "FEI FALCON II (4k x 4k)"
-		case reFalconI.MatchString(value):
-			return "FEI FALCON I (4k x 4k)"
-		}
-	}
-
-	log.Printf("value %v is not in enum %s!", value, namePDBx)
-	// if not found in enum list and it's a funding organisation, put a certain string
-	if namePDBx == "_pdbx_audit_support.funding_organization" {
-		return "Other government"
-	}
-	// // if no match and enum contains option for OTHER, choose it
-	// if converterUtils.SliceContains(enumFromEMDB, "OTHER") || converterUtils.SliceContains(enumFromPDBx, "OTHER"){
-	// 	return "OTHER"
-	// }
-
-	return strings.ToUpper(value)
-}
-
-func checkValue(dataItem converterUtils.PDBxItem, value string, jsonKey string, unitsOSCEM string) string {
-
-	//now based on the found struct implement range matching, units matching and enum matching
-	if dataItem.ValueType == "int" || dataItem.ValueType == "float" {
-		namePDBx := dataItem.CategoryID + "." + dataItem.Name
-		// in OSCEM defocus is negative value and overfocus is positive. In PDBx it's vice versa if string starts with  minus, ut it off, othersise add a - prefix
-		if namePDBx == "_em_imaging.nominal_defocus_min" || namePDBx == "_em_imaging.calibrated_defocus_min" || namePDBx == "_em_imaging.nominal_defocus_max" || namePDBx == "_em_imaging.calibrated_defocus_max" {
-			// change the sign negative to positive and vice versa
-			if value[0] == 45 {
-				value = value[1:]
-			} else {
-				value = "-" + value
-			}
-		}
-		validatedRange, err := validateRange(value, dataItem, unitsOSCEM, jsonKey)
-		if err != nil {
-			errorNumeric := fmt.Sprintf("JSON value %s not numeric, but supposed to be", value)
-			if err.Error() == errorNumeric {
-				return "?"
-			} else {
-				// FIXME when units conversion is implemented, handle this error
-				log.Println(err.Error())
-			}
-		}
-		if !validatedRange {
-			log.Printf("Value %s of property %s is not in range of [ %s, %s ]!\n", value, jsonKey, dataItem.RangeMin, dataItem.RangeMax)
-		}
-	} else if dataItem.ValueType == "yyyy-mm-dd" {
-		value = validateDateIsRFC3339(value)
-	} else if len(dataItem.EnumValues) > 0 || len(dataItem.PDBxEnumValues) > 0 {
-		value = validateEnum(value, dataItem)
-	}
-
-	if strings.Contains(value, " ") {
-		value = fmt.Sprintf("'%s' ", value) // if name contains whitespaces enclose it in single quotes
-	} else {
-		value = fmt.Sprintf("%s ", value) // take value as is
-	}
-	return value
-}
-
 func getOrderCategories(parsedCategories []string, mmCIFCategories []string) []string {
 	var order []string
 	allCategories := append(parsedCategories, mmCIFCategories...)
-	sort.Slice(allCategories, func(i, j int) bool {
-		return allCategories[i] < allCategories[j]
-	})
+	slices.Sort(allCategories)
 	// sort based on the pre-defined (administrative, polymer related entities, ligand (non-polymer) related instances, and structure level description)
-	for _, category := range converterUtils.PDBxCategoriesOrder {
+	for _, category := range cU.PDBxCategoriesOrder {
 		category = "_" + category
-		if converterUtils.SliceContains(allCategories, category) {
+		if cU.SliceContains(allCategories, category) {
 			order = append(order, category)
 		}
 	}
 	// add the rest not atom-related in some order (it can be random)
 	for _, c := range allCategories {
-		if !converterUtils.SliceContains(order, c) && !converterUtils.SliceContains(converterUtils.PDBxCategoriesOrderAtom, c[1:]) && !(len(c) > 5 && c[0:5] == "data_") {
+		if !cU.SliceContains(order, c) && !cU.SliceContains(cU.PDBxCategoriesOrderAtom, c[1:]) && !(len(c) > 5 && c[0:5] == "data_") {
 			order = append(order, c)
 		}
 
 	}
 	// add atoms categories
-	for _, category := range converterUtils.PDBxCategoriesOrderAtom {
+	for _, category := range cU.PDBxCategoriesOrderAtom {
 		category = "_" + category
-		if converterUtils.SliceContains(allCategories, category) {
+		if cU.SliceContains(allCategories, category) {
 			order = append(order, category)
 		}
 	}
 	// append the rest of "unparsed" categories that were inside their own "data_" containers
 	for i := range mmCIFCategories {
-		if mmCIFCategories[i][0:5] == "data_" {
-			order = append(order, mmCIFCategories[i])
+		if len(mmCIFCategories[i]) > 5 {
+			if mmCIFCategories[i][0:5] == "data_" {
+				order = append(order, mmCIFCategories[i])
+			}
 		}
 	}
 	return order
 }
 
-func parseMmCIF(dictFile *os.File) (string, map[string]string, error) {
-	scanner := bufio.NewScanner(dictFile)
+func parseMmCIF(dictR io.Reader) (string, map[string]string, error) {
+	scanner := bufio.NewScanner(dictR)
 
 	var longestData uint32 = 0
 	var dataToStr = make(map[string]string, 0)
@@ -440,7 +213,8 @@ func parseMmCIF(dictFile *os.File) (string, map[string]string, error) {
 	return longestNameData, mmCIFfieldsMain, nil
 }
 
-func LoopDataEntry(scanner *bufio.Scanner, category string) (map[string]string, string, uint32, string) { // return the categories, same as long string, the length and next category
+// return the categories, same as long string, the length and next category
+func LoopDataEntry(scanner *bufio.Scanner, category string) (map[string]string, string, uint32, string) {
 	var l uint32 = 0
 	var str strings.Builder
 	var asText strings.Builder
@@ -479,13 +253,13 @@ func LoopDataEntry(scanner *bufio.Scanner, category string) (map[string]string, 
 
 // CreteMetadataCif creates and mmCIF file content as a string. This function only converts metadata in the required format.
 // Meant for EMDB depositions where no other existing mmCIF is available
-func CreteMetadataCif(nameMapper map[string]string, PDBxItems map[string][]converterUtils.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string) (string, error) {
+func CreteMetadataCif(nameMapper map[string]string, PDBxItems map[string][]cU.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string) (string, error) {
 	return createCifText("data_myID", map[string]string{}, nameMapper, PDBxItems, valuesMap, OSCEMunits)
 }
 
 // Given an mmCIF file create a new one with added scientific Metadata.
 // Meant for PDB depositions
-func SupplementCoordinatesFromFile(nameMapper map[string]string, PDBxItems map[string][]converterUtils.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string, mmCIFpath *os.File) (string, error) {
+func SupplementCoordinates(nameMapper map[string]string, PDBxItems map[string][]cU.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string, mmCIFpath io.Reader) (string, error) {
 	var dataID string
 	var mmCIFvalues map[string]string
 
@@ -500,26 +274,18 @@ func SupplementCoordinatesFromFile(nameMapper map[string]string, PDBxItems map[s
 
 // Given an mmCIF file path, open it and create a new one with added scientific Metadata.
 // Meant for PDB depositions
-func SupplementCoordinatesFromPath(nameMapper map[string]string, PDBxItems map[string][]converterUtils.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string, mmCIFpath string) (string, error) {
-	var dataID string
-	var mmCIFvalues map[string]string
-	dictFile, err := os.Open(mmCIFpath)
+func SupplementCoordinatesFromPath(nameMapper map[string]string, PDBxItems map[string][]cU.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string, mmCIFpath string) (string, error) {
+	dictR, err := os.Open(mmCIFpath)
 	if err != nil {
 		errorString := fmt.Sprintf("mmCIF file %s does not exist!", mmCIFpath)
 		return "", errors.New(errorString)
 	}
-	defer dictFile.Close()
+	defer dictR.Close()
 
-	name, categories, err := parseMmCIF(dictFile)
-	if err != nil {
-		return "", err
-	}
-	dataID = name
-	mmCIFvalues = categories
-	return createCifText(dataID, mmCIFvalues, nameMapper, PDBxItems, valuesMap, OSCEMunits)
+	return SupplementCoordinates(nameMapper, PDBxItems, valuesMap, OSCEMunits, dictR)
 }
 
-func createCifText(dataName string, mmCIFCategories map[string]string, nameMapper map[string]string, PDBxItems map[string][]converterUtils.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string) (string, error) {
+func createCifText(dataName string, mmCIFCategories map[string]string, nameMapper map[string]string, PDBxItems map[string][]cU.PDBxItem, valuesMap map[string][]string, OSCEMunits map[string][]string) (string, error) {
 	// keeps track of values from JSON that have already been mapped to the PDBx properties
 	var str strings.Builder
 	str.WriteString(dataName + "\n#\n") //write the data Identifier in the header
@@ -528,7 +294,7 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 	for k := range PDBxItems {
 		parsedCategories = append(parsedCategories, k)
 	}
-	allCategories := getOrderCategories(parsedCategories, converterUtils.GetKeys(mmCIFCategories))
+	allCategories := getOrderCategories(parsedCategories, cU.GetKeys(mmCIFCategories))
 
 	for _, category := range allCategories {
 		var duplicatedFlag bool = false
@@ -571,6 +337,14 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 				valuesStr.WriteString("")
 				var isRelevantID bool
 				for _, dataItem := range catDI {
+					if cifDIlen != 0 && size == cifDIlen {
+						log.Printf(
+							"Category %s exists both in metadata from JSON files and in existing mmCIF file! "+
+								"Since your data has many instances of this category, I can't automatically complement "+
+								"the instance with data_item %s",
+							dataItem.CategoryID, dataItem.Name)
+						continue
+					}
 					jsonKey, err := getKeyByValue(dataItem.CategoryID+"."+dataItem.Name, nameMapper)
 					if err != nil {
 						// it is _id property -> check if we need it go through all data items and see if it's a parent somewhere!
@@ -591,6 +365,10 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 
 				for v := range size {
 					for _, dataItem := range catDI {
+						if cifDIlen != 0 && size == cifDIlen {
+							// do not complement
+							continue
+						}
 						jsonKey, err := getKeyByValue(dataItem.CategoryID+"."+dataItem.Name, nameMapper)
 						if err != nil {
 							if isRelevantID {
@@ -612,7 +390,7 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 							}
 						}
 					}
-					if len(cifDIs) != 0 && size == cifDIlen {
+					if cifDIlen != 0 && size == cifDIlen {
 						for _, k := range keys {
 							var value string
 							if strings.Contains(cifDIs[k][v], " ") {
@@ -625,7 +403,7 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 					}
 					valuesStr.WriteString("\n")
 				}
-				if len(cifDIs) != 0 && size == cifDIlen {
+				if cifDIlen != 0 && size == cifDIlen {
 					for _, k := range keys { // write in the same order as extracted from the mmCIF
 						fmt.Fprintf(&str, "%s\n", k)
 					}
@@ -633,7 +411,7 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 				str.WriteString(valuesStr.String())
 				str.WriteString("#\n")
 			case size == 1:
-				l := getLongestPDBxItem(catDI, converterUtils.GetKeys(cifDIs)) + 5
+				l := getLongestPDBxItem(catDI, cU.GetKeys(cifDIs)) + 5
 				var isRelevantID bool
 				for _, dataItem := range catDI {
 					jsonKey, err := getKeyByValue(dataItem.CategoryID+"."+dataItem.Name, nameMapper)
@@ -643,7 +421,7 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 						if isRelevantID {
 							// remove the ID entry key from list of parsed from mmCIF we will use one from metadata
 							// delete(cifDIs, dataItem.CategoryID+"."+dataItem.Name)
-							// keys = converterUtils.DeleteElementFromList(keys, dataItem.CategoryID+"."+dataItem.Name)
+							// keys = cU.DeleteElementFromList(keys, dataItem.CategoryID+"."+dataItem.Name)
 							if _, ok := cifDIs[dataItem.CategoryID+"."+dataItem.Name]; !ok {
 								formatString := fmt.Sprintf("%%-%ds", l)
 								fmt.Fprintf(&str, formatString, dataItem.CategoryID+"."+dataItem.Name)
@@ -660,12 +438,13 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 					if jsonValue, ok := valuesMap[jsonKey]; ok {
 						// remove that key from list of parsed from mmCIF we will use one from metadata
 						// delete(cifDIs, dataItem.CategoryID+"."+dataItem.Name)
-						// keys = converterUtils.DeleteElementFromList(keys, dataItem.CategoryID+"."+dataItem.Name)
+						// keys = cU.DeleteElementFromList(keys, dataItem.CategoryID+"."+dataItem.Name)
 						if _, ok := cifDIs[dataItem.CategoryID+"."+dataItem.Name]; !ok {
 							formatString := fmt.Sprintf("%%-%ds", l)
 							fmt.Fprintf(&str, formatString, dataItem.CategoryID+"."+dataItem.Name)
 							if unit, ok := OSCEMunits[jsonKey]; ok {
-								valueString := checkValue(dataItem, jsonValue[0], jsonKey, unit[0]) // the 0th element, because it's the case where only one value is present
+								// the 0th element, because it's the case where only one value is present
+								valueString := checkValue(dataItem, jsonValue[0], jsonKey, unit[0])
 								fmt.Fprintf(&str, "%s", valueString)
 							} else {
 								// values that have no units definition in OSCEM
@@ -697,7 +476,8 @@ func createCifText(dataName string, mmCIFCategories map[string]string, nameMappe
 				}
 				str.WriteString("#\n")
 			default:
-				// Based on conversion table, the correspondence in naming between OSCEM and PDBx exist. But for this PDBx data category not OSCEM properties are used in this JSON file.
+				// Based on conversion table, the correspondence in naming between OSCEM and PDBx exist.
+				// But for this PDBx data category not OSCEM properties are used in this JSON file.
 				continue
 			}
 		}
